requirements.txt
astroid==2.3.1
certifi==2018.4.16
colorama==0.4.1
cycler==0.10.0
Django==2.0.4
django-rest-auth==0.9.3
djangorestframework==3.9.0
isort==4.3.21
kiwisolver==1.0.1
lazy-object-proxy==1.4.2
matplotlib==3.0.1
mccabe==0.6.1
mysqlclient==1.3.12
numpy==1.15.4
pandas==0.23.4
pipenv==2018.7.1
pylint==2.4.2
pymongo==3.10.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.4
six==1.12.0
typed-ast==1.4.0
virtualenv==16.1.0
virtualenv-clone==0.3.0
virtualenvwrapper-win==1.2.5
wrapt==1.11.2
